Branch jake increment del1

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -4,12 +4,17 @@ import static java.util.Objects.requireNonNull;
 
 import java.util.List;
 
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.Region;
+import javafx.stage.Stage;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
+import seedu.address.ui.ConfirmDeleteWindow;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
@@ -24,11 +29,36 @@ public class DeleteCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
+    public static final String CONFIRM_DELETE = "Are you sure you want to delete this person?";
+    public static final String TITLE = "CollabSync - Confirm Delete";
+    public static final String FEEDBACK_TO_USER_CANCELLED_DELETE = "Deletion canceled.";
+    public static final String FEEDBACK_TO_USER_UNKNOWN_ERROR = "Unexpected error occured. Please try again";
 
     private final Index targetIndex;
 
+    // This flag is used to mock a situation where users enter "confirm" when deleting a person
+    // True for JUnit testing, but otherwise, false for default boolean values
+    private boolean isConfirmedForTesting;
+
+    /**
+     * Default constructor for creating objects of DeleteCommand.
+     *
+     * @param targetIndex The index of the person to be deleted.
+     */
     public DeleteCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
+    }
+
+    /**
+     * Constructor that is only used for JUnit Testing in {@code DeleteCommandTest}, where we are required
+     * to mock a situation where users enter "confirm" when deleting a person.
+     *
+     * @param targetIndex The index to be deleted.
+     * @param isConfirmedForTesting Value passed in is always {@code true} for JUnit testing only.
+     */
+    public DeleteCommand(Index targetIndex, boolean isConfirmedForTesting) {
+        this.targetIndex = targetIndex;
+        this.isConfirmedForTesting = isConfirmedForTesting;
     }
 
     @Override
@@ -41,8 +71,40 @@ public class DeleteCommand extends Command {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deletePerson(personToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
+
+        if (isConfirmedForTesting) {
+            model.deletePerson(personToDelete);
+            return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
+        } else {
+            try {
+                // Load the FXML file and get the controller
+                FXMLLoader loader = new FXMLLoader(getClass().getResource("/view/ConfirmDeleteWindow.fxml"));
+                Region dialogRoot = loader.load();
+                ConfirmDeleteWindow deleteController = loader.getController();
+
+                // Set the confirmation message on the controller
+                deleteController.setConfirmationMessage(CONFIRM_DELETE);
+
+                // Show the confirmation dialog box in a new window
+                Stage dialogStage = new Stage();
+                dialogStage.setTitle(TITLE);
+                dialogStage.setScene(new Scene(dialogRoot));
+
+                dialogStage.showAndWait(); // Wait for user's response
+
+                // Based on the user's inputs to the confirmation dialog box, handle the deletion logic
+                if (deleteController.isConfirmed()) {
+                    model.deletePerson(personToDelete);
+                    return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS,
+                            Messages.format(personToDelete)));
+                } else {
+                    return new CommandResult(FEEDBACK_TO_USER_CANCELLED_DELETE);
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new CommandResult(FEEDBACK_TO_USER_UNKNOWN_ERROR);
+            }
+        }
     }
 
     @Override
